Load owner's asset content manager
